+++ packages/store/src/internal/internals.ts
@@ -216,13 +216,16 @@ export function propGetter(paths: string[], config: NgxsConfig) {
 export function buildGraph(stateClasses: StateClassInternal[]): StateKeyGraph {
   const findName = (stateClass: StateClassInternal) => {
     const meta = stateClasses.find(g => g === stateClass);
-    if (!meta) {
+
+    // Caretaker note: we have still left the `typeof` condition in order to avoid
+    // creating a breaking change for projects that still use the View Engine.
+    if ((typeof ngDevMode === 'undefined' || ngDevMode) && !meta) {
       throw new Error(
         `Child state not found: ${stateClass}. \r\nYou may have forgotten to add states to module`
       );
     }
 
-    return meta[META_KEY]!.name!;
+    return meta![META_KEY]!.name!;
   };
 
   return stateClasses.reduce<StateKeyGraph>(
@@ -332,7 +335,9 @@ export function topologicalSort(graph: StateKeyGraph): string[] {
     visited[name] = true;
 
     graph[name].forEach((dep: string) => {
-      if (ancestors.indexOf(dep) >= 0) {
+      // Caretaker note: we have still left the `typeof` condition in order to avoid
+      // creating a breaking change for projects that still use the View Engine.
+      if ((typeof ngDevMode === 'undefined' || ngDevMode) && ancestors.indexOf(dep) >= 0) {
         throw new Error(
           `Circular dependency '${dep}' is required by '${name}': ${ancestors.join(' -> ')}`
         );